move analogy run to init
